Enhance flexibility for local development
