+++ tests/govtool-frontend/playwright/tests/wallet.bootstrap.ts
@@ -1,5 +1,5 @@
 import { adaHolderWallets, dRepWallets } from "@constants/staticWallets";
-import { setAllureEpic } from "@helpers/allure";
+import { setAllureEpic, setAllureStory } from "@helpers/allure";
 import { pollTransaction } from "@helpers/transaction";
 import { expect, test as setup } from "@playwright/test";
 import kuberService from "@services/kuberService";
@@ -9,6 +9,7 @@ setup.describe.configure({ mode: "serial", timeout: environments.txTimeOut });
 
 setup.beforeEach(async () => {
   await setAllureEpic("Setup");
+  await setAllureStory("Wallet bootstrap");
 });
 
 setup("Initialize static wallets", async () => {
